+++ lib/impersonate/browser/chrome.ts
@@ -3,18 +3,19 @@ import { deepMerge, parseHeaders } from '../util'
 import type { ImpersonateConfig, VariantImpersonateConfig } from '../types'
 
 export enum ChromeBrowser {
-  Chrome133 = 'chrome133a',
-  Chrome134 = 'chrome134',
-  Chrome = 'chrome134',
+  Chrome136 = 'chrome136',
+  Chrome142 = 'chrome142',
+  Chrome143 = 'chrome143',
+  Chrome = 'chrome143',
 }
 
 export const DEFAULT_CHROME_FINGERPRINT = {
   ja3: '771,4865-4866-4867-49195-49199-49196-49200-52393-52392-49171-49172-156-157-47-53,0-5-10-11-13-16-18-23-27-35-43-45-51-17613-65037-65281,4588-29-23-24,0',
-  ja4: 't13d1516h2_1301,1302,1303,c02b,c02f,c02c,c030,cca9,cca8,c013,c014,009c,009d,002f,0035_0033,000b,0023,002b,000d,0000,ff01,0005,002d,000a,0010,0012,001b,fe0d,44cd,0017_0403,0804,0401,0503,0805,0501,0806,0601',
+  ja4: 't13d1516h2_1301,1302,1303,c02b,c02f,c02c,c030,cca9,cca8,c013,c014,009c,009d,002f,0035_44cd,fe0d,0012,000a,0005,0017,ff01,001b,002b,0010,000b,000d,0000,0023,002d,0033_0403,0804,0401,0503,0805,0501,0806,0601',
   akami: '1:65536;2:0;4:6291456;6:262144|15663105|0|m,a,s,p',
 }
 
-export const DEFAULT_CHROME_VERSION = '134'
+export const DEFAULT_CHROME_VERSION = '143'
 
 export function getChromeConfig(config?: VariantImpersonateConfig) {
   const version = config?.version ?? DEFAULT_CHROME_VERSION
@@ -23,7 +24,7 @@ export function getChromeConfig(config?: VariantImpersonateConfig) {
     {
       compressed: true,
       headers: parseHeaders([
-        `sec-ch-ua: "Not(A:Brand";v="99", "Google Chrome";v="${version}", "Chromium";v="${version}"`,
+        `sec-ch-ua: "Google Chrome";v="${version}", "Chromium";v="${version}", "Not A(Brand";v="24"`,
         'sec-ch-ua-mobile: ?0',
         'sec-ch-ua-platform: "Windows"',
         'Upgrade-Insecure-Requests: 1',
@@ -52,6 +53,7 @@ export function getChromeConfig(config?: VariantImpersonateConfig) {
 
 export const CHROME_BROWSER_CONFIGS: Record<ChromeBrowser, ImpersonateConfig> =
   {
-    [ChromeBrowser.Chrome133]: getChromeConfig({ version: '133' }),
-    [ChromeBrowser.Chrome134]: getChromeConfig(),
+    [ChromeBrowser.Chrome136]: getChromeConfig({ version: '136' }),
+    [ChromeBrowser.Chrome142]: getChromeConfig({ version: '142' }),
+    [ChromeBrowser.Chrome143]: getChromeConfig(),
   }