+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.1.0] - Unreleased
+
+### Added
+- Build for Node.js 24 Krypton
+
+### Changed
+- Updated default impersonate profiles to:
+  - Chrome 143
+  - Edge 143
+  - Firefox 144
+  - Safari 18.6
+- Updated curl-impersonate to v1.2.5
+
+### Fixed
+- Build on macOS due to outdated zlib version
+
 ## [5.0.3] - 2025-04-03
 
 ### Added
@@ -449,6 +465,7 @@ Special Thanks to [@koskokos2](https://github.com/koskokos2) for their contribut
 - Improved code style, started using prettier
 ## [1.2.0] - 2017-08-28
 
+[5.1.0]: https://github.com/andrewmackrodt/node-libcurl-ja3/compare/v5.0.3...HEAD
 [5.0.3]: https://github.com/andrewmackrodt/node-libcurl-ja3/compare/v5.0.2...v5.0.3
 [5.0.2]: https://github.com/andrewmackrodt/node-libcurl-ja3/compare/v5.0.1...v5.0.2
 [5.0.1]: https://github.com/andrewmackrodt/node-libcurl-ja3/compare/v5.0.0...v5.0.1